Use `f32` instead of `f64`

+++ benches/haversine.rs
@@ -8,15 +8,15 @@ use criterion::Criterion;
 use aeroscore::haversine::haversine_distance;
 
 struct Point {
-    latitude: f64,
-    longitude: f64,
+    latitude: f32,
+    longitude: f32,
 }
 
 impl aeroscore::Point for Point {
-    fn latitude(&self) -> f64 {
+    fn latitude(&self) -> f32 {
         self.latitude
     }
-    fn longitude(&self) -> f64 {
+    fn longitude(&self) -> f32 {
         self.longitude
     }
     fn altitude(&self) -> i16 {
